Move print logic to cli tools
