+++ crates/wasm-metadata/Cargo.toml
@@ -13,7 +13,6 @@ workspace = true
 [dependencies]
 clap = { workspace = true, optional = true }
 anyhow = { workspace = true }
-comfy-table = { workspace = true }
 wasmparser = { workspace = true, features = ['component-model', 'hash-collections'] }
 wasm-encoder = { workspace = true, features = ['component-model'] }
 indexmap = { workspace = true, features = ["serde"] }
